Allows normalized Levenshtein distance metric
Adds a new flag, -m or --method which can alter the comparison method
used to detect duplicates. By default, this remains the original
Levenshtein metric. Additionally though, a user can specify the
'normalized-levenshtein' metric.

When specified, this interprets the distance parameter as a number
between zero and one and is used for setting the maximum allowed
normalized Levenshtein distance between non-unique inputs.